move kernel version check to its own pkg
