+++ pkg/kernel/kernel_utils_test.go
@@ -1,4 +1,4 @@
-package utils
+package kernel
 
 import (
 	"testing"
